requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cli-helpers==2.3.0
click==8.1.3
configobj==5.0.6
idna==3.4
pendulum==2.1.2
pgcli==3.5.0
pgspecial==2.0.1
prompt-toolkit==3.0.32
psycopg==3.1.4
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytzdata==2020.1
requests==2.28.1
setproctitle==1.3.2
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
